requirements.txt
asgiref==3.8.1
Brotli==1.1.0
build==1.2.2.post1
cffi==1.17.1
chardet==5.2.0
click==8.1.8
cssselect2==0.7.0
Django==5.1.5
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
docopt==0.6.2
fonttools==4.55.5
num2words==0.5.14
packaging==24.2
pillow==11.1.0
pip-tools==7.4.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dotenv==1.0.1
reportlab==4.2.5
setuptools==75.8.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==63.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
